initial setup
In this directory, I want to create a project named Speaker Assistant. Please initialize git repo first, and important rule: I want all my prompts that change anything in this project, to be a git commit messages. in a form: <Summary of changes> <my prompt>. Meaning this prompt should be a first commit named sth like: initial setup <text of this prompt>. Always update readme to be up to date with new requirements and instructions. Here I want to do experiment that I store history of the prompts and the codebase behind, At some point I'll try out different approaches on whether I have to version the codebase, or just a history of the prompts is enough. Note that if I ask you for something and you have a followup questions before you make changes, do not commit yet, and include my answers in the commit message. commit after i accept the changes